The ESP32 allows larger message sizes

+++ src/ConfigurableFirmata.cpp
@@ -817,7 +817,7 @@ void FirmataClass::setPinState(byte pin, byte state)
 void FirmataClass::systemReset(void)
 {
   resetting = true;
-  byte i;
+  int i;
 
   waitForData = 0; // this flag says the next serial input will be data
   executeMultiByteCommand = 0; // execute this after getting multi-byte data
